+++ ExampleProject/index.ios.js
@@ -1,5 +1,0 @@
-'use strict';
-import { AppRegistry } from 'react-native';
-import ExampleProject from './ExampleProject.js';
-
-AppRegistry.registerComponent('ExampleProject', () => ExampleProject);