+++ Starter/ios/Starter/AppDelegate.m
@@ -22,7 +22,7 @@
   RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];
   RCTRootView *rootView = [[RCTRootView alloc] initWithBridge:bridge
                                                    moduleName:@"Starter"
-                                            initialProperties:nil];
+                                            initialProperties:nil];  
 
   rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];
 
@@ -31,7 +31,6 @@
   rootViewController.view = rootView;
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
-  
   // Add CleverTap Account ID and Account Token in your .plist file)
   // initialize CleverTap
 #ifdef DEBUG